refactor: split slides into separate components
