+++ src/components/LogoWatermark.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const LogoWatermark = () => (
+  <svg className="logo-watermark" viewBox="0 0 2011 1566" xmlns="http://www.w3.org/2000/svg">
+    <path d="M2011 1566L1201 173L506 1391L301 1390L1105 0L900 2L0 1563L605 1561L1205 527L1710 1392L1503 1389L1204 869L1105 1045L1405 1562L2011 1566Z" fill="#00872b"/>
+  </svg>
+);
+
+export default LogoWatermark;
